Split (foreground) symbol placement code out of symbol_bucket (background buffer generation+serialization). Use CollisionBox struct array to pass collision information from background to foreground. Strip out information we're no longer using.
